+++ Core/Core.cs
@@ -1,24 +1,91 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Diagnostics;
+using System.Globalization;
+using System.Linq;
 using Eve.Plugin;
+using Eve.Ref;
+using Eve.Types;
 
 namespace Eve.Core {
-	[Serializable]
-	internal class Core : IPlugin {
-		public string Id => Guid.NewGuid().ToString();
+	public class Core : IPlugin {
 		public string Name => "Core";
-		public string Version => "2.0";
+		public string Author => "SemiViral";
+		public string Version => "3.0.2";
+		public string Id => Guid.NewGuid().ToString();
 		public bool TerminateRequestRecieved { get; private set; }
+		public Dictionary<string, string> Commands => new Dictionary<string, string> {
+			["eval"] = "(<expression>) — evaluates given mathematical expression.",
+			["join"] = "(<channel>) — joins specified channel."
+		}; 
 		public PluginStatus Status { get; private set; } = PluginStatus.Stopped;
-		public Dictionary<string, string> Definition => null;
 
 		public void ProcessEnded() {
-			if (!TerminateRequestRecieved) {}
+			if (!TerminateRequestRecieved) { }
 		}
 
 		public void OnChannelMessage(object source, ChannelMessageEventArgs e) {
 			Status = PluginStatus.Processing;
+
+			if (!Commands.Keys.Contains(e.SplitArgs[1].ToLower())) {
+				Status = PluginStatus.Stopped;
+				return;
+			}
+
+			PluginEventArgs responseEvent = new PluginEventArgs {
+				MessageType = PluginEventMessageType.Message
+			};
+
+			PluginChannelMessageResponse response = new PluginChannelMessageResponse(Protocols.PRIVMSG, e.Recipient,
+				string.Empty);
+
+			switch (e.SplitArgs[1]) {
+				case "eval":
+					if (e.SplitArgs.Count < 3) {
+						response.Message = "Not enough parameters.";
+						responseEvent.Result = response;
+						DoCallback(responseEvent);
+						Status = PluginStatus.Stopped;
+						break;
+					}
+
+					Status = PluginStatus.Running;
+					string evalArgs = e.SplitArgs.Count > 3 ?
+						e.SplitArgs[2] + e.SplitArgs[3] : e.SplitArgs[2];
+
+					try {
+						response.Message = new Calculator().Evaluate(evalArgs).ToString(CultureInfo.CurrentCulture);
+						responseEvent.Result = response;
+						DoCallback(responseEvent);
+					} catch (Exception ex) {
+						response.Message = ex.Message;
+						responseEvent.Result = response;
+						DoCallback(responseEvent);
+					}
+					break;
+				case "join":
+					if (User.Current.Access > 1)
+						response.Message = "Insufficient permissions.";
+					else if (e.SplitArgs.Count < 3)
+						response.Message = "Insufficient parameters. Type 'eve help join' to view command's help index.";
+					else if (!e.SplitArgs[2].StartsWith("#"))
+						response.Message = "Channel name must start with '#'.";
+					else if (Channel.Get(e.SplitArgs[2].ToLower()) != null)
+						response.Message = "I'm already in that channel.";
+
+					if (!string.IsNullOrEmpty(response.Message)) {
+						responseEvent.Result = response;
+						DoCallback(responseEvent);
+						return;
+					}
+
+					response.Target = string.Empty;
+					response.Message = e.SplitArgs[2];
+					response.Protocol = Protocols.JOIN;
+					responseEvent.Result = response;
+					DoCallback(responseEvent);
+					break;
+			}
 
 			Status = PluginStatus.Stopped;
 		}
@@ -63,31 +130,14 @@ namespace Eve.Core {
 
 	//public class Join : IPlugin {
 	//	public Dictionary<string, string> Def => new Dictionary<string, string> {
-	//		["join"] = "(<channel>) — joins specified channel."
+	//		
 	//	};
 
 	//	public ChannelMessage OnChannelMessage(ChannelMessage c, PassableMutableObject v) {
 	//		if (!c.SplitArgs[1].CaseEquals(Def.Keys.First()))
 	//			return null;
 
-	//		if (v.CurrentUser.Access > 1)
-	//			c.message = "Insufficient permissions.";
-	//		else if (string.IsNullOrEmpty(c.SplitArgs[2]))
-	//			c.message = "Insufficient parameters. Type 'eve help join' to view ccommand's help index.";
-	//		else if (!c.SplitArgs[2].StartsWith("#"))
-	//			c.message = "Channel c._Argsument must be a proper channel name (i.e. starts with '#').";
-	//		else if (v.GetChannel(c.SplitArgs[2].ToLower()) != null)
-	//			c.message = "I'm already in that channel.";
 
-	//		if (!string.IsNullOrEmpty(c.message)) {
-	//			c.Type = Protocols.Privmsg;
-	//			return c;
-	//		}
-
-	//		c.Target = string.Empty;
-	//		c.message = c.SplitArgs[2];
-	//		c.Type = Protocols.Join;
-	//		return c;
 	//	}
 	//}
 
@@ -180,9 +230,9 @@ namespace Eve.Core {
 	//		if (c.SplitArgs.Count < 4) c.message = "Insufficient parameters. Type 'eve help message' to view correct usage.";
 	//		else who = c.SplitArgs[2].ToLower();
 
-	//		if (v.GetUser(who) == null)
+	//		if (v.Get(who) == null)
 	//			c.message = "User does not exist in database.";
-	//		else if (!v.GetUser(who).AddMessage(new message {
+	//		else if (!v.Get(who).AddMessage(new message {
 	//			Sender = c.Nickname,
 	//			Contents = Regex.Escape(c.SplitArgs[3]),
 	//			Date = DateTime.UtcNow
@@ -235,7 +285,7 @@ namespace Eve.Core {
 	//	}
 	//}
 
-	//public class GetUser : IPlugin {
+	//public class Get : IPlugin {
 	//	public Dictionary<string, string> Def => new Dictionary<string, string> {
 	//		["user"] = "(<user>) — returns stored nickname and access level of specified user."
 	//	};
@@ -246,14 +296,14 @@ namespace Eve.Core {
 
 	//		if (c.SplitArgs.Count < 3)
 	//			c.message = "Insufficient parameters. Type 'eve help user' to view command's help index.";
-	//		else if (v.GetUser(c.SplitArgs[2].ToLower()) == null)
+	//		else if (v.Get(c.SplitArgs[2].ToLower()) == null)
 	//			c.message = $"User {c.SplitArgs[2]} does not exist in database.";
 
 	//		if (!string.IsNullOrEmpty(c.message))
 	//			return c;
 
 	//		string who = c.SplitArgs[2].ToLower();
-	//		c.message = $"{v.GetUser(who).Realname}({v.GetUser(who).Access})";
+	//		c.message = $"{v.Get(who).Realname}({v.Get(who).Access})";
 
 	//		return c;
 	//	}
@@ -270,13 +320,13 @@ namespace Eve.Core {
 
 	//		if (c.SplitArgs.Count < 3)
 	//			c.message = "Insufficient parameters. Type 'eve help seen' to view correct usage.";
-	//		else if (v.GetUser(c.SplitArgs[2].ToLower()) == null)
+	//		else if (v.Get(c.SplitArgs[2].ToLower()) == null)
 	//			c.message = "User does not exist in database.";
 
 	//		if (!string.IsNullOrEmpty(c.message))
 	//			return c;
 
-	//		User u = v.GetUser(c.SplitArgs[2].ToLower());
+	//		User u = v.Get(c.SplitArgs[2].ToLower());
 	//		c.message = $"{u.Realname} was last seen on: {u.Seen} (UTC)";
 
 	//		return c;
@@ -297,7 +347,7 @@ namespace Eve.Core {
 
 	//		if (c.SplitArgs.Count < 4)
 	//			c.message = "Insuffient parameters. Type 'eve help setaccess' to view command's help index.";
-	//		else if (v.GetUser(c.SplitArgs[2]) == null)
+	//		else if (v.Get(c.SplitArgs[2]) == null)
 	//			c.message = "User does not exist in database.";
 	//		else if (!int.TryParse(c.SplitArgs[3], out i))
 	//			c.message = "Invalid access parameter.";
@@ -306,7 +356,7 @@ namespace Eve.Core {
 	//			c.message = "Invalid access setting. Please use a number between 1 and 9.";
 	//		else if (v.CurrentUser.Access > 1)
 	//			c.message = "Insufficient permissions. Only users with an access level of 1 or 0 can promote.";
-	//		else if (!v.GetUser(c.SplitArgs[2]).SetAccess(i))
+	//		else if (!v.Get(c.SplitArgs[2]).SetAccess(i))
 	//			c.message = "Error occured, aborting operation.";
 
 	//		if (!string.IsNullOrEmpty(c.message))