+++ Eve/Plugin/Wrapper.cs
@@ -1,18 +1,35 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel.Design;
 using System.Diagnostics;
 
 namespace Eve.Plugin {
-	internal class PluginWrapper {
+	internal class PluginWrapper : MarshalByRefObject {
+		public static Dictionary<string, string> Commands = new Dictionary<string, string>();
 		public PluginHost PluginHost;
 
-		private static void PluginsCallback(object source, PluginEventArgs e) {
+		private void PluginsCallback(object source, PluginEventArgs e) {
 			switch (e.MessageType) {
 				case PluginEventMessageType.Message:
-					Writer.Log(e.ResultMessage, EventLogEntryType.Information);
+					if (e.Result is PluginChannelMessageResponse) {
+						PluginChannelMessageResponse response = (PluginChannelMessageResponse)e.Result;
+						Writer.SendData(response.Protocol, $"{response.Target} {response.Message}");
+						break;
+					}
+
+					Writer.Log(e.Result.ToString(), EventLogEntryType.Information);
 					break;
 				case PluginEventMessageType.EventLog:
 					break;
 				case PluginEventMessageType.Action:
+					if (e.EventAction.ActionToTake == PluginActionType.AddCommand) {
+						if (!(e.Result is KeyValuePair<string, string>)) break;
+
+						var temp = (KeyValuePair<string, string>)e.Result;
+						Commands.Add(temp.Key, temp.Value);
+					} else if (false) {
+						/* later conditionals */
+					}
 					break;
 				default:
 					throw new ArgumentOutOfRangeException();